Refactor and comment
