+++ pkg/utils/overriding/test-fixtures/patches/command-and-env/result.yaml
