+++ pkg/utils/overriding/test-fixtures/patches/kubernetes-component-union/original.yaml
