+++ pkg/utils/overriding/test-fixtures/patches/kubernetes-component-union/result.yaml
